requirements.txt
cryptography==43.0.0
Kivy==2.3.0
kivymd==1.2.0
python_dateutil==2.9.0.post0
db-sqlite3